test/index.html

<!DOCTYPE html>

<html lang=en>

<meta charset=utf-8>

<title>snuggsi ツ Test Suite</title>

<!-- Testing Format http://web-platform-tests.org
<script src=testharness.js></script>
<script src=testharnessreport.js></script>
-->

<script src=https://unpkg.com/snuggsi></script>

<script>
//test (test => {
//  assert_true (true, 'Test stub works!')
//}, 'Descriptive test name')
</script>

<link rel=match href="${2:URL of match}">
<link rel=help  href=//path/to/specification>


<body role=application>

<header role=banner>${banner}</header>

<nav>${nav}</nav>

<main id=content>
${content}
</main>

<footer role=contentinfo>${contentinfo}</footer>

<output>${log}</output>
